Wire: Ignore bus arbitration errors
This is a workaround for noise on the bus causing bus arbitration
errors, which then cause the Wire library to lock up and the system to
reset (which causes the "display freeze" bug, since with the u8glib I2c
implementation, the system would not lock up,  but only I2c transactions
would fail). This fix breaks multi-master operation, and thus only works
because we know there will be no other masters on the bus.