Wire: Add timeouts to all loops
This should prevent lockups due to the hardware behaving in unexpected
ways. Previously, a reset was executed on arbitration failures, but it
seems there were more ways to get the hardware into an unexpected state
that caused to not process requests (probably thinking there is another
master on the bus).

This commit does not prevent the hardware issues themselves, but at
least detects them through a timeout and reset the hardware when it
happens.